requirements.txt
flask
flask-sqlalchemy
flask-migrate
flask_session
flask_login 
psycopg2-binary
werkzeug
python-dotenv
requests